backend/requirements.txt
fastapi
uvicorn
httpx
phonenumbers
pydantic
python-multipart
python-dotenv
slowapi
